# src/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "MySQL Speaker",
    "description": "API Documentation of MySQL Speaker"
  },
  "servers": [
    {
      "url": "https://mysqlspeaker.online/",
      "description": ""
    }
  ],
  "paths": {
...
    "/api/database": {
      "post": {
        "tags": [
          "Create"
        ],
        "description": "Endpoint for creating a database.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized access, no token provided or invalid token"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "dbName": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Delete"
        ],
        "description": "Endpoint for deleting a database.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized access, no token provided or invalid token"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "dbName": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/table": {
      "post": {
        "tags": [
          "Create"
        ],
        "description": "Endpoint for creating a table.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized access, no token provided or invalid token"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "dbName": {
                    "example": "any"
                  },
                  "table": {
                    "example": "any"
                  },
                  "columns": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Delete"
        ],
        "description": "Endpoint for deleting a table.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized access, no token provided or invalid token"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "dbName": {
                    "example": "any"
                  },
                  "table": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/databases": {
      "get": {
        "tags": [
          "Read"
        ],
        "description": "Endpoint to obtain all databases and their tables.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized access, no token provided or invalid token"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/data/query": {
      "post": {
        "tags": [
          "Read"
        ],
        "description": "Endpoint to query specific data from a table.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized access, no token provided or invalid token"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "dbName": {
                    "example": "any"
                  },
                  "table": {
                    "example": "any"
                  },
                  "select": {
                    "example": "any"
                  },
                  "where": {
                    "example": "any"
                  },
                  "groupBy": {
                    "example": "any"
                  },
                  "orderBy": {
                    "example": "any"
                  },
                  "orderDirection": {
                    "example": "any"
                  },
                  "offset": {
                    "example": "any"
                  },
                  "limit": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/data": {
      "put": {
        "tags": [
          "Update"
        ],
        "description": "Endpoint to update individual data.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized access, no token provided or invalid token"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "dbName": {
                    "example": "any"
                  },
                  "table": {
                    "example": "any"
                  },
                  "data": {
                    "example": "any"
                  },
                  "where": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Update"
        ],
        "description": "Endpoint to insert new row data.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized access, no token provided or invalid token"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "dbName": {
                    "example": "any"
                  },
                  "table": {
                    "example": "any"
                  },
                  "values": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Delete"
        ],
        "description": "Endpoint for deleting specific data.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized access, no token provided or invalid token"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "dbName": {
                    "example": "any"
                  },
                  "table": {
                    "example": "any"
                  },
                  "where": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\""
      }
    },
...
  }
}